Finish Rating

+++ src/main/java/com/th/bean/MovieRating.java
@@ -6,7 +6,7 @@ import java.io.Serializable;
 
 /**
  * <p>
- * 
+ *
  * </p>
  *
  * @author TanHaooo
@@ -20,10 +20,9 @@ public class MovieRating implements Serializable {
 
     private Integer movieId;
 
-    private Integer rating;
+    private Float rating;
 
     private String timestamp;
-
 
     public Integer getUserId() {
         return userId;
@@ -41,11 +40,11 @@ public class MovieRating implements Serializable {
         this.movieId = movieId;
     }
 
-    public Integer getRating() {
+    public Float getRating() {
         return rating;
     }
 
-    public void setRating(Integer rating) {
+    public void setRating(Float rating) {
         this.rating = rating;
     }
